move ConditionalWrapper to separate file
